python/requirements.txt
absl-py
aiodns
aiohttp
anyio
apache-beam
argon2-cffi
astunparse
async-generator
async-timeout
attrs
autopep8
avro-python3
Babel
backcall
bleach
brotlipy
cachetools
cchardet
certifi
cffi
chardet
click
colorama
crcmod
cycler
decorator
defusedxml
dill
docopt
entrypoints
fastavro
fasteners
Flask
flatbuffers
future
gast
gevent
google-api-core
google-api-python-client
google-apitools
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigtable
google-cloud-core
google-cloud-datastore
google-cloud-dlp
google-cloud-language
google-cloud-pubsub
google-cloud-spanner
google-cloud-videointelligence
google-cloud-vision
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-gcp
grpcio-tools
gunicorn
h5py
hdfs
httplib2
idna
importlib-resources
imutils
ipykernel
ipython
ipython-genutils
itsdangerous
jedi
Jinja2
joblib
json5
jsonschema
jupyter-client
jupyter-core
jupyter-server
jupyterlab
jupyterlab-pygments
jupyterlab-server
keras-nightly
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
multidict
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
oauth2client
oauthlib
opencv-contrib-python
opt-einsum
packaging
pandas
pandocfilters
parso
pickleshare
Pillow
prometheus-client
promise
prompt-toolkit
proto-plus
protobuf
pyarrow
pyasn1
pyasn1-modules
pycares
pycodestyle
pycparser
pydot
Pygments
pymongo
pyparsing
pyrsistent
python-dateutil
python-ffmpeg-video-streaming
pytz
pywin32
pywinpty
pyzmq
requests
requests-futures
requests-oauthlib
requests-unixsocket
rsa
scikit-learn
scipy
Send2Trash
simplejpeg
six
sniffio
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-metadata
tensorflow-serving-api
tensorflow-transform
termcolor
terminado
testpath
tfds-nightly
tflite-runtime
pycoral
tfx-bsl
threadpoolctl
toml
tornado
tqdm
traitlets
typing-extensions
uritemplate
urllib3
wcwidth
webencodings
websocket-client
Werkzeug
wrapt
yarl
zipp
zope.event
zope.interface
